# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Contacts API Documentation",
    "version": "1.0.0",
    "description": "A comprehensive API for managing contact information, built with Express and MongoDB.",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
      "Contact": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The auto-generated id of the contact"
          },
          "firstName": {
            "type": "string",
            "description": "The first name of the contact"
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the contact"
          },
          "email": {
            "type": "string",
            "description": "The email address of the contact"
          },
          "phoneNumber": {
            "type": "string",
            "description": "The phone number of the contact"
          },
          "company": {
            "type": "string",
            "description": "The company name"
          },
          "role": {
            "type": "string",
            "description": "The job role/title"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date the contact was added"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date the contact was last updated"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/contacts": {
      "get": {
        "summary": "Returns the list of all the contacts",
        "tags": [
          "Contacts"
        ],
        "responses": {
          "200": {
            "description": "The list of the contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Contact"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new contact",
        "tags": [
          "Contacts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The contact was successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or duplicate email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/contacts/{id}": {
      "get": {
        "summary": "Get the contact by id",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The contact id"
          }
        ],
        "responses": {
          "200": {
            "description": "The contact description by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "The contact was not found"
          }
        }
      },
      "put": {
        "summary": "Update the contact by the id",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The contact id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The contact was updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "The contact was not found"
          }
        }
      },
      "delete": {
        "summary": "Remove the contact by id",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The contact id"
          }
        ],
        "responses": {
          "200": {
            "description": "The contact was deleted"
          },
          "404": {
            "description": "The contact was not found"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Contacts",
      "description": "The contacts managing API"
    }
  ]
}